Create 16s sequencing code
All the code needed to run all the analyses for figures 2-4 in the manuscript. This code uses Supplementary data file 3-5, which can be used to make a phyloseq object. If you downloaded the raw data from NCBI and are following this code your results may be a little different because we included additional samples that were in the run for the removal of ASV calling, quality filtering and the removal of contaminants etc. Raw data from these samples have not been uploaded because they are being analyzed for a follow-up study.